feat: PR #132 Phase 1 - Complete Int64→Int migration and critical fixes
Implements first 8 fixes from PR #132 code review (Priority 0 items #4-6).

## Type System Migration (Int64 → Int)
- Update RestoreImageRecord.fileSize: Int64 → Int
- Update XcodeVersionRecord.fileSize: Int64? → Int?
- Remove unnecessary Int(record.fileSize) conversions in RecordBuilder
- Update IPSWFetcher, AppleDBFetcher, IPSWVersion to use Int
- Rename FieldValueExtensions.int64Value → intValue (+ 2 usage updates)

## Boolean Helper Usage
- Replace manual `.int64(bool ? 1 : 0)` with `.boolean(bool)` in RecordBuilder (4 locations)
- Move FieldValue.boolean() to extension for SwiftLint type_contents_order compliance

## Critical Bug Fixes
- Fix force unwrap crash risk in XcodeReleasesFetcher.toDate (use nil-coalescing)
- Fix duplicate BushelCloudKitService initialization in SyncEngine (reuse single instance)
- Remove silent error swallowing (try? → direct assignment)

## Build Status
✅ Clean build (3.81s) with no errors

Related to: #132 (items #1, #4, #5, #6, #10)

+++ Examples/Bushel/Sources/BushelImages/DataSources/AppleDB/AppleDBFetcher.swift
@@ -109,7 +109,7 @@ struct AppleDBFetcher: Sendable {
             buildNumber: build,
             releaseDate: releaseDate ?? Date(), // Fallback to current date
             downloadURL: link.url,
-            fileSize: Int64(ipswSource.size ?? 0),
+            fileSize: ipswSource.size ?? 0,
             sha256Hash: ipswSource.hashes?.sha2_256 ?? "",
             sha1Hash: ipswSource.hashes?.sha1 ?? "",
             isSigned: isSigned,
